Move the header to a separate component JC-703
