+++ apps/react/src/App.module.css
@@ -1,0 +1,6 @@
+.main {
+  display: flex;
+  height: 100vh;
+  max-height: 100vh;
+  width: 100%;
+}
