+++ apps/react/src/App.tsx
@@ -4,6 +4,9 @@ import { BrowserRouter } from 'react-router-dom';
 
 import { RootRouter } from './routes/RootRouter';
 import { store } from './store';
+import { Header } from './components/Header';
+
+import styles from './App.module.css';
 
 /** The main app component. */
 export const App: FC = () => (
@@ -11,7 +14,10 @@ export const App: FC = () => (
 		<BrowserRouter>
 			<div>
 				<Suspense fallback={<div>Brrr... here should be your loader component</div>}>
-					<RootRouter />
+					<div className={styles.main}>
+						<Header/>
+						<RootRouter />
+					</div>
 				</Suspense>
 			</div>
 		</BrowserRouter>
