+++ apps/react/src/components/NavigationList/NavigationList.tsx
@@ -7,7 +7,7 @@ import ListItemText from '@mui/material/ListItemText';
 import InboxIcon from '@mui/icons-material/MoveToInbox';
 import MailIcon from '@mui/icons-material/Mail';
 
-import { NavigationProps } from '../../utils/navigationProps';
+import { NavigationProps } from '@js-camp/react/features/genres/utils/navigationProps';
 
 type Props = {
