--- apps/react/src/features/genres/components/GenreLayout/index.ts
@@ -1,1 +1,0 @@
-export { GenreLayout } from './GenreLayout';
