+++ apps/react/src/features/genres/pages/GenrePage/GenrePage.module.css
@@ -1,38 +1,35 @@
-.main {
+.layout {
   display: flex;
-  height: 100vh;
-  max-height: 100vh;
+  gap: var(--space-8);
+  padding: 100px var(--space-6) 0 var(--space-6);
+  transition: margin 225ms cubic-bezier(0.4, 0, 0.6, 1);
+
+  /* Set a negative margin when side-menu is closed */
+  margin-left: calc(-1 * var(--sidebar-width));
+  height: 100%;
   width: 100%;
 }
 
-.main__app-bar {
-  transition: margin 225ms cubic-bezier(0.4, 0, 0.6, 1), width 225ms cubic-bezier(0.4, 0, 0.6, 1);
-  position: fixed;
+.layout_open {
+  transition: margin 195ms cubic-bezier(0.4, 0, 0.2, 1);
+  margin-left: 0;
 }
 
-.main__app-bar_open {
-  width: calc(100% - var(--sidebar-width));
-  margin-left: var(--sidebar-width);
-  transition: margin 195ms cubic-bezier(0.4, 0, 0.2, 1), width 195ms cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-.main__toolbar {
-  height: var(--header-height);
-}
-
-.main__toolbar-icon {
-  margin-right: var(--space-4);
-}
-
-.main__drawer {
-  width: var(--sidebar-width);
+.layout__sidebar {
+  width: var(--list-width);
   flex-shrink: 0;
 }
 
-.main__drawer-header {
+.layout__empty {
   display: flex;
   align-items: center;
-  padding: 0 var(--space-2);
-  justify-content: flex-end;
-  height: var(--header-height);
+  justify-content: center;
+  flex-grow: 1;
+  margin: 0 auto;
+}
+
+.layout__button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
